the consumer not lost messages, automatic ack selected. Documentation added
